WIP upgrades – very much not working yet!
See https://github.com/JustinBusschau/php-govtalk/pull/3 for a guide to
what some of the remaining changes will need to look like. To maintain
Guzzle library cross-compatability etc., another requirement for this
update will be for the upstream `php-govtalk` to be swapped over to its
v1.x (if updated in situ) or to the `thebiggive` fork.